Test1

+++ mulawAPI.h
@@ -20,7 +20,7 @@
 
 #define magnitude(sample) ((sample < 0) ? -sample : sample) 
 
-#define signum(num) ((num > 0)) 
+#define signum(num) (unsigned int)((num > 0)) 
 
 
 // Functions 
@@ -29,7 +29,7 @@
 
  
 
-int codeword_compression(unsigned int sample_magnitude, int sign); 
+int codeword_compression(unsigned int sample_magnitude, unsigned int sign); 
 
 int codeword_decompression(int codeWord);
